+++ tasks/w-cmd.js
@@ -37,7 +37,8 @@ var
                 options: {
                     '-h, --help': 'print usage information',
                     '-v, --version': 'print yyl version',
-                    '-p, --path': 'show the yyl command local path'
+                    '-p, --path': 'show the yyl command local path',
+                    '--logLevel': 'log level'
                 }
             });
         },